task 3

+++ 0x01-python_async_function/3-tasks.py
@@ -9,6 +9,6 @@ from typing import Awaitable, Any, Callable, Coroutine
 wait_random = __import__('0-basic_async_syntax').wait_random
 
 
-def task_wait_random(max_delay: int) -> Callable[[float], int]:
+def task_wait_random(max_delay: int) -> Callable[[int], float]:
     '''defining the function'''
     return asyncio.ensure_future(wait_random(max_delay))
